Before moving vars
